+++ .gitignore
@@ -105,7 +105,7 @@ celerybeat.pid
 *.sage.py
 
 # Environments
-.env
+server/.env
 .venv
 env/
 venv/
@@ -130,7 +130,7 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
-/.env
+/server/.env
 
 # temporary files
 :memory
