+++ server/server/views/utils/jwt.py
@@ -1,5 +1,5 @@
 from datetime import datetime, timedelta
-from ...config import Config
+from server.config import Config
 import jwt
